### 変更内容 productDAOクラスを2クラスに分割
### 背景
メソッドが1クラス内に2個定義されていたため、
保守性を考慮し1クラス1メソッドへリファクタリング。
